added quick1k script
Added script I'm using to quickly play 500 rounds of 2 games to see if
we've done anything horrendous

if you're seeing this, don't use it as a way to definitively tell if an
engine is better! The time controls are way too short and it doesn't
play enough games. This is just for a quick test.

bench 2357736